1.1.3.8 Preview

+++ scripts/_xkeen/02_install/07_install_register/01_register_xray.sh
@@ -68,9 +68,22 @@ register_xray_initd() {
     backup_path="${backups_dir}/${s24xray_filename}"
     backup_paths="${backups_dir}/${s99start_filename}"
     script_file="${xinstall_dir}/07_install_register/04_register_init.sh"
-    variables_to_extract="name_client name_policy table_id table_mark port_dns ipv4_proxy ipv4_exclude ipv6_proxy ipv6_exclude port_donor port_exclude start_attempts check_fd arm64_fd other_fd delay_fd"
+    variables_to_extract="name_client name_policy table_id table_mark port_dns ipv4_proxy ipv4_exclude ipv6_proxy ipv6_exclude port_donor port_exclude start_attempts start_auto check_fd arm64_fd other_fd delay_fd"
     temp_file=$(mktemp)
     start_delay_value=20 # Задержка автозапуска по умолчанию
+    autostart="on" # Автозапуск по умолчанию
+
+    if [ -f "${start_file}" ]; then
+        mv "${start_file}" "${backup_paths}"
+        if grep -q "^autostart=" "${backup_paths}"; then
+            autostart_value=$(grep "^autostart=" "${backup_paths}" | head -n 1 | cut -d'=' -f2)
+            [ -n "$autostart_value" ] && autostart="$autostart_value"
+        fi
+        if grep -q "^start_delay=" "${backup_paths}"; then
+            start_delay_value=$(grep "^start_delay=" "${backup_paths}" | head -n 1 | cut -d'=' -f2)
+            [ -n "$start_delay_value" ] && start_delay="$start_delay_value"
+        fi
+    fi
 
     if [ ! -e "${initd_file}" ]; then
         cp "${script_file}" "${initd_file}"
@@ -78,14 +91,6 @@ register_xray_initd() {
         echo -e "  Файл автозапуска ${yellow}создан и обновлен${reset}"
     else
         mv "${initd_file}" "${backup_path}"
-
-        if [ -f "${start_file}" ]; then
-            mv "${start_file}" "${backup_paths}"
-            if grep -q "^start_delay=" "${backup_paths}"; then
-                start_delay_value=$(grep "^start_delay=" "${backup_paths}" | head -n 1 | cut -d'=' -f2)
-            fi
-        fi
-
         cat "${script_file}" > "${initd_file}"
 
         for var in $variables_to_extract; do
@@ -105,19 +110,17 @@ register_xray_initd() {
 }
 
 register_autostart() {
-    rm -f "${initd_dir}/S99xkeenrestart"
+    [ -f "/opt/etc/init.d/S99xkeenrestart" ] && rm "/opt/etc/init.d/S99xkeenrestart"
     cat << EOF > "${initd_dir}/S99xkeenstart"
 #!/bin/sh
 #
-autostart="on"
+autostart=${autostart_value}
 start_delay=${start_delay_value}
 #
 log_info_router() {
     logger -p notice -t "XKeen" "\$1"
 }
-if [ -f "/opt/etc/init.d/S99xkeenrestart" ]; then
-  rm "/opt/etc/init.d/S99xkeenrestart"
-fi
+[ -f "/opt/etc/init.d/S99xkeenrestart" ] && rm "/opt/etc/init.d/S99xkeenrestart"
 if [ "\${autostart}" = "on" ]; then
     HOST="ya.ru"
     while true; do
